Implementation of kqueue event machine
Currently not tested on target system, so following patches may be
bugfixing that.

Interface changed slightly in that it now provides universal values
EVENT_READ and EVENT_WRITE. Each of those correspond to appropriate
implementation value.

Some entries in EM_result enum changed and one was added:

  * EM_ERROR_EPOLL_CREATE -> EM_ERROR_EVENT_CREATE_QUEUE
  * EM_ERROR_EPOLL_CTL -> EM_ERROR_EVENT_CTL
  * EM_ERROR_EPOLL_WAIT -> EM_ERROR_EVENT_WAIT
  * Added EM_ERROR_FCNTL.

There is now also macro HAVE_PIPE2 that is turned on on Linux, since it
has pipe2() system call. Additional fcntl() calls are necessary on
systems which doesn't support it.